Moved getSKU to utils folder
